Tht server 78 그날의 폴링 주제를 선택한 유저 리스트 조회 (#82)
* add : 데일리 주제 미선택 커스텀 exception

* add : Test Fixture

* test : 메인화면 유저 리스트 추출 테스트

* feat : 유저 메인 메칭 화면 리스트 페이징 조회 (커버링 인덱싱)

+++ api/src/main/java/com/tht/api/app/entity/user/UserInterests.java
@@ -28,16 +28,16 @@ public class UserInterests {
     private String userUuid;
 
     @Column(name = "interest_idx")
-    private Long interestIdx;
+    private Integer interestIdx;
 
     @Builder(access = AccessLevel.PRIVATE)
-    public UserInterests(final Long idx, final String userUuid, final Long interestIdx) {
+    public UserInterests(final Long idx, final String userUuid, final Integer interestIdx) {
         this.idx = idx;
         this.userUuid = userUuid;
         this.interestIdx = interestIdx;
     }
 
-    public static UserInterests create(final String userUuid, final Long interestIdx) {
+    public static UserInterests create(final String userUuid, final Integer interestIdx) {
         return UserInterests.builder()
             .userUuid(userUuid)
             .interestIdx(interestIdx)
